game setup & handling Login command

+++ core/src/main/scala/xox/core/codecs/ServerCommandCodec.scala
@@ -11,18 +11,18 @@ object ServerCommandCodec {
   lazy val codec: Codec[ServerCommand] = Codec(encoder, decoder)
 
   lazy val encoder: Encoder[ServerCommand] =
-    (byte ~ commandEncoder.encodeOnly)
+    (uint8 ~ commandEncoder.encodeOnly)
       .xmapc { case (_, command) => command } (command => commandCode(command) -> command)
 
   lazy val decoder: Decoder[ServerCommand] =
-    byte.flatMap(commandDecoder)
+    uint8.flatMap(commandDecoder)
 
   private lazy val commandEncoder =
     selectedEncoder[ServerCommand](cmd => Err(s"Cannot encode server command: $cmd")) {
       case _: Login => loginCodec.upcast
     }
 
-  private def commandDecoder(code: Byte): Decoder[ServerCommand] =
+  private def commandDecoder(code: Int): Decoder[ServerCommand] =
     code match {
       case 1       => loginCodec
       case unknown => fail(Err(s"Unknown server command code: $unknown"))
@@ -30,7 +30,7 @@ object ServerCommandCodec {
 
   private lazy val loginCodec = ascii.as[Login]
 
-  private def commandCode(command: ServerCommand): Byte =
+  private def commandCode(command: ServerCommand): Int =
     command match {
       case _: Login => 1
     }
